+++ src/nex/repl.clj
@@ -41,7 +41,7 @@
   ["class" "enum" "deferred" "declare" "feature" "inherit" "end" "do" "if" "then" "else" "elseif"
    "when" "from" "until" "invariant" "variant" "require" "ensure"
    "let" "create" "convert" "to" "fn" "function" "and" "or" "old" "this" "note"
-   "with" "import" "intern" "private" "raise" "rescue" "retry" "spawn" "select" "timeout" "repeat" "across" "case" "of"
+   "with" "import" "intern" "private" "raise" "rescue" "retry" "assert" "spawn" "select" "timeout" "repeat" "across" "case" "of"
    "true" "false" "nil"
    ;; strictly 'result' is not a keyword, but a pre-defined variable name.
    "result"])
@@ -481,6 +481,7 @@
       "pre" :pre
       "post" :post
       "invariant" :invariant
+      "assert" :assert
       s)))
 
 (defn- repl-cmd-help [ctx input]